+++ app/src/main/java/ca/pkay/rcloneexplorer/AboutActivity.java
@@ -1,25 +1,26 @@
 package ca.pkay.rcloneexplorer;
 
+import android.app.ActivityManager;
 import android.content.Intent;
+import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
+import android.preference.PreferenceManager;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.mikepenz.aboutlibraries.Libs;
-import com.mikepenz.aboutlibraries.LibsBuilder;
 
 public class AboutActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        applyTheme();
         setContentView(R.layout.activity_about);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
@@ -66,6 +67,19 @@ public class AboutActivity extends AppCompatActivity {
         });
     }
 
+    private void applyTheme() {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        int customPrimaryColor = sharedPreferences.getInt("pref_key_color_primary", -1);
+        int customAccentColor = sharedPreferences.getInt("pref_key_color_accent", -1);
+        getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
+        getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
+
+        // set recents app color to the primary color
+        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
+        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
+        setTaskDescription(taskDesc);
+    }
+
     @Override
     public boolean onSupportNavigateUp() {
         onBackPressed();
@@ -78,13 +92,8 @@ public class AboutActivity extends AppCompatActivity {
     }
 
     private void showOpenSourceLibraries() {
-        new LibsBuilder()
-                .withActivityStyle(Libs.ActivityStyle.LIGHT_DARK_TOOLBAR)
-                .withActivityTitle(getString(R.string.credits_libraries))
-                .withAutoDetect(false)
-                .withLibraries()
-                .withExcludedLibraries()
-                .start(this);
+        Intent librariesIntent = new Intent(this, AboutLibsActivity.class);
+        startActivity(librariesIntent);
     }
 
     private void openAppGitHubLink() {