+++ app/src/main/java/ca/pkay/rcloneexplorer/MainActivity.java
@@ -3,15 +3,20 @@ package ca.pkay.rcloneexplorer;
 import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.app.ActivityManager;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
@@ -57,6 +62,7 @@ public class MainActivity extends AppCompatActivity
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        applyTheme();
         context = this;
         setContentView(R.layout.activity_main);
         Toolbar toolbar = findViewById(R.id.toolbar);
@@ -94,6 +100,19 @@ public class MainActivity extends AppCompatActivity
             startRemotesFragment();
         }
 
+    }
+
+    private void applyTheme() {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        int customPrimaryColor = sharedPreferences.getInt("pref_key_color_primary", -1);
+        int customAccentColor = sharedPreferences.getInt("pref_key_color_accent", -1);
+        getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
+        getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
+
+        // set recents app color to the primary color
+        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
+        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
+        setTaskDescription(taskDesc);
     }
 
     @Override
@@ -150,6 +169,10 @@ public class MainActivity extends AppCompatActivity
                 } else {
                     importConfigFile();
                 }
+                break;
+            case R.id.nav_settings:
+                Intent settingsIntent = new Intent(this, SettingsActivity.class);
+                startActivity(settingsIntent);
                 break;
             case R.id.nav_about:
                 Intent aboutIntent = new Intent(this, AboutActivity.class);